+++ tests/pb/sample_read_with_binary
