+++ tests/pb/test_pb_write.py
@@ -2,7 +2,7 @@ import os
 import unittest
 
 import time
-import slspb
+import aliyun_log_pb as slspb
 import gzip
 import pickle
 import warnings
