+++ tests/pb/test_suite.py
@@ -1,9 +1,9 @@
 import unittest
-from tests.test_pb_parse import TestParsePB
-from tests.test_pb_write import TestWritePB
-from tests.test_pb_misc import TestPBMisc
-from tests.test_pb_compare import TestWritePB as TestPBCompare
-from tests.test_pb_stress import TestWritePB as TestWritePBStress
+from .test_pb_parse import TestParsePB
+from .test_pb_write import TestWritePB
+from .test_pb_misc import TestPBMisc
+from .test_pb_compare import TestWritePB as TestPBCompare
+from .test_pb_stress import TestWritePB as TestWritePBStress
 
 if __name__ == '__main__':
     suite = unittest.TestSuite()
